+++ pi_copterVS/pi_copterVS/Stabilization.h
@@ -24,56 +24,56 @@ class StabilizationClass{
 
 private:
 	
-	void max_speed_limiter(float &x, float &y);
-	void dist2speed(float &x, float &y);
-	void speed2dist(float &x, float &y);
+	void max_speed_limiter(double &x, double &y);
+	void dist2speed(double &x, double &y);
+	void speed2dist(double &x, double &y);
 
 	
-	float SPEED_Z_CF, SPEED_XY_CF;
+	double SPEED_Z_CF, SPEED_XY_CF;
 
-	float mc_z,mc_x,mc_y,d_speedX,d_speedY;
+	double mc_z,mc_x,mc_y,d_speedX,d_speedY;
 	float dist2speed_XY;
 	float alt2speedZ;
-	float accxy_stab(float dist, float maxA, float timeL);
-	float accxy_stab_rep(float speed, float maxA, float timeL);
-	//float throttle;
+	double accxy_stab(double dist, double maxA, double timeL);
+	double accxy_stab_rep(double speed, double maxA, double timeL);
+	//double throttle;
 #define STAB_PIDS 3
 	AP_PID pids[STAB_PIDS];
-	float max_z_integrator;
+	double max_z_integrator;
 
-	float needXR, needYR, needXV, needYV;
+	double needXR, needYR, needXV, needYV;
 
-	void set_acc_xy_speed_kp(const float f){ pids[SPEED_X_SPEED].kP(f);	pids[SPEED_Y_SPEED].kP(f); }
-	void set_acc_xy_speed_kI(const float f){ pids[SPEED_X_SPEED].kI(f);	pids[SPEED_Y_SPEED].kI(f); }
-	void set_acc_xy_speed_kD(const float f){ pids[SPEED_X_SPEED].kD(f,3);	pids[SPEED_Y_SPEED].kD(f,3); }
-	void set_acc_xy_speed_imax(const float f){ pids[SPEED_X_SPEED].imax(-f,f);	pids[SPEED_Y_SPEED].imax(-f,f); }
+	void set_acc_xy_speed_kp(const double f){ pids[SPEED_X_SPEED].kP(f);	pids[SPEED_Y_SPEED].kP(f); }
+	void set_acc_xy_speed_kI(const double f){ pids[SPEED_X_SPEED].kI(f);	pids[SPEED_Y_SPEED].kI(f); }
+	void set_acc_xy_speed_kD(const double f){ pids[SPEED_X_SPEED].kD(f,3);	pids[SPEED_Y_SPEED].kD(f,3); }
+	void set_acc_xy_speed_imax(const double f){ pids[SPEED_X_SPEED].imax(-f,f);	pids[SPEED_Y_SPEED].imax(-f,f); }
 	
 	
 public:
 	void setMaxAng();
 	void setMinMaxI_Thr();
 	void setNeedPos2Home();
-	void add2NeedPos(float speedX, float speedY, float dt);
-	void setNeedPos(float x, float y);
-	void setNeedLoc(long lat, long lon, float &x, float &y);
-	void fromLoc2Pos(long lat, long lon, float &x, float &y);
-	float get_dist2goal();
+	void add2NeedPos(double speedX, double speedY, double dt);
+	void setNeedPos(double x, double y);
+	void setNeedLoc(long lat, long lon, double &x, double &y);
+	void fromLoc2Pos(long lat, long lon, double &x, double &y);
+	double get_dist2goal();
 	//void set_XY_2_GPS_XY();
 	void  resset_z();
 	void  resset_xy_integrator();
-	//float getAltitude() { return sZ; }
-	float getSpeed_Z(const float dist){
+	//double getAltitude() { return sZ; }
+	double getSpeed_Z(const double dist){
 		return constrain(dist*alt2speedZ, max_speedZ_M, max_speedZ_P);
 	}
 
-	float getDist_Z(float speed){
+	double getDist_Z(double speed){
 		return speed/alt2speedZ;
 	}
 
-	float getSpeed_XY(float dist){		
+	double getSpeed_XY(double dist){		
 		return dist* dist2speed_XY;
 	}
-	float getDist_XY(float speed){	
+	double getDist_XY(double speed){	
 		return speed/ dist2speed_XY;
 	}
 
@@ -85,7 +85,7 @@ public:
 
 
 	float max_speedZ_P,max_speedZ_M,max_speed_xy;
-	float Z();
+	double Z();
 	void XY(float &xF, float&yF);
 
 	void init();
